ai-sub2/requirement.txt
absl-py==0.7.1
astor==0.8.0
bothub==0.1.30
bothub-cli==0.1.20
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dialogflow==0.6.0
Flask==0.12.2
gast==0.2.2
google-api-core==1.14.2
google-auth==1.6.3
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.23.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
JPype1==0.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
konlpy==0.5.1
Markdown==3.1.1
MarkupSafe==1.1.1
nltk==3.4.3
numpy==1.16.3
pandas==0.24.2
pathspec==0.5.9
prompt-toolkit==2.0.9
protobuf==3.9.1
pyaml==19.4.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pyee==6.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
rouge==0.3.2
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
slack==0.0.2
slackclient==1.3.2
slackeventsapi==2.1.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.15.5
wrapt==1.11.2
zmq==0.0.0
